requirements.txt
langchain
fastapi 
uvicorn
openai-whisper
python-dotenv
langchain_community
langchain-core
langchain-openai
openai
langchain-groq
python-multipart